Fix clippy lints

+++ src/encoder.rs
@@ -44,8 +44,8 @@ pub fn delta_encode<R: Read, W: Write>(source: R, target: R, patch: W) -> io::Re
 
 #[inline]
 fn create_instructions(source: &[u8], target: &[u8]) -> Vec<u8> {
-    debug_assert!(source.len() <= CHUNK_SIZE as usize);
-    debug_assert!(target.len() <= CHUNK_SIZE as usize);
+    debug_assert!(source.len() <= CHUNK_SIZE);
+    debug_assert!(target.len() <= CHUNK_SIZE);
     let lcs = Lcs::new(source, target).subsequence();
     let mut bytes: Vec<u8> = Vec::with_capacity(max(source.len(), target.len()) + lcs.len());
     let mut source_index: usize = 0;
